+++ sh_document_management/__manifest__.py
@@ -9,8 +9,8 @@
     "category": "Document Management",
     "summary": "View  Document Module, Filter Document,Document Management,Manage Doc Group By Document, Get Document Details, Document Manage, Manage Directory Document Odoo",
     "description": """This module helps to manage documents easily in odoo. You can create the directory and manage directory wise documents. Users can filter the documents by visible directory & my directory. You can easily group by the documents by directory, custom date & created by. You can easily add custom filters/groups of documents. From the menu bar, the user can see directory tags & document tags. Using a search bar you can search documents details easily. You can download the document from the files in the directory. You can see related documents from the sub-directory.""",
-    "version": "0.2.0",
-    "depends": ["base", "mail","web"],
+    "version": "0.3.0",
+    "depends": ["base", "mail", "web", "portal"],
     "application": True,
     "post_init_hook": "post_init_hook",
     "data": [
@@ -21,6 +21,7 @@
         'views/document_tags_views.xml',
         'views/directory_tags_views.xml',
         'views/document_directory_views.xml',
+        'views/portal_templates.xml',
         'data/ir_cron_data.xml',
         'data/mail_template_data.xml',
     ],